services/vector_store/requirements.txt
faiss-cpu
sentence-transformers
grpcio
grpcio-tools
psycopg2-binary
